+++ src/main/java/br/ufpb/dcx/Loja/Loja.java
@@ -4,8 +4,8 @@ import java.util.Collection;
 
 public interface Loja {
 
-    public boolean cadastraContato(String nome, int cpf, int celular);
-    public Collection<Cliente> pesquisarClientes(int cpf, int celular);
+    public boolean cadastraContato(String nome, long cpf, long celular);
+    public Collection<Cliente> pesquisarClientes(long cpf, long celular);
     public boolean removeContato(String nome);
     public boolean adicionarCarro(String modelo, int ano, double quilometragem, String placa);
     public boolean removerCarro(String placa);
